% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{create_db}
\alias{create_db}
\title{Create a database}
\usage{
create_db(fn, tz_step, n, tz_origin = 0, size = 8)
}
\arguments{
\item{fn}{file name}

\item{tz_step}{step of time series data in seconds}

\item{n}{number of steps to store}

\item{tz_origin}{origin of time steps in seconds}

\item{size}{size in bits of the written number data}
}
\value{
Called for the bi-product of creating the file
}
\description{
Create a round robin database for a series of numeric data collected at regular time intervals
}
