% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{read_db}
\alias{read_db}
\title{Read data from the database}
\usage{
read_db(fn, strt, fnsh, tz = "UTC")
}
\arguments{
\item{fn}{file name}

\item{strt}{POSIXct (or interger) object for the start of the period}

\item{fnsh}{POSIXct (or interger) object for the end of the period}

\item{tz}{time zone used for the output}
}
\value{
an xts object containing any data in the time period
}
\description{
Reads data for a specfied time window from the database
}
