% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrdb-package.R
\docType{package}
\name{rrdb-package}
\alias{rrdb}
\alias{rrdb-package}
\title{rrdb: A Simple Round Robin Database}
\description{
A simple round robin database for storing fixed length histories of timeseries data. Inspired by whisperDB, but not feature or structure compatable.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://waternumbers.github.io/rrdb/}
  \item \url{https://github.com/waternumbers/rrdb}
  \item Report bugs at \url{https://github.com/waternumbers/rrdb/issues}
}

}
\author{
\strong{Maintainer}: Paul Smith \email{paul@waternumbers.co.uk} (\href{https://orcid.org/0000-0002-0034-3412}{ORCID})

}
\keyword{internal}
