% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{update_db}
\alias{update_db}
\title{Update the data in a database}
\usage{
update_db(fn, D, check = TRUE)
}
\arguments{
\item{fn}{file name}

\item{D}{the xts data object to write to the file}

\item{check}{should the age of the overwritten data be checked}
}
\value{
Called for the bi-product of updating the data
}
\description{
Adds data to the database overwriting existing data in a round robin fashion
}
